+++ mappings/Rossler.h
@@ -22,13 +22,13 @@ class Rossler : public Mapping<T> {
 public:
 
     Rossler(const T &coeff_1, const T &coeff_2, const T &coeff_3, const T &coeff_4,
-            const T &coeff_5, int composition_constant);
+            const T &coeff_5, size_t composition_constant);
 
 };
 
 template<typename T>
 Rossler<T>::Rossler(const T &coeff_1, const T &coeff_2, const T &coeff_3, const T &coeff_4,
-                    const T &coeff_5, int composition_constant) : Mapping<T>(composition_constant), coeff_1(coeff_1), coeff_2(coeff_2),
+                    const T &coeff_5, size_t composition_constant) : Mapping<T>(composition_constant), coeff_1(coeff_1), coeff_2(coeff_2),
                                                                   coeff_3(coeff_3), coeff_4(coeff_4),
                                                                   coeff_5(coeff_5)  {}
